Install texlive via installation script
I wanted to get independed of the destribution packages. Therefore I
decided to install texlive via an installtion script that you can
download from tug.org. I use the latest frozen version from 2022. I
think that is good enough for me, but I might update later, once a new
version is frozen.

My hope to do it his way, that I will be able to install on other base
images (e.g. alpine). I know for alpine the path has to change slightly,
but that should not be such a big deal. I would have gone with alpine,
if pandoc was available from the package repository of alpine linux, but
it is not (at least at the moment).