Remove apt-get upgrade
This causes problems when there are packages that actually can be
upgraded. One problem is, that the command tells you that additional
diskspace is needed and asks you if you want to continue. If you don't
agree, the command aborts. It might be possible to circumvent this with
a command line parameter. But my bigger concern is, that updating any
packages might break the rest of the build.